+++ src/lib.rs
@@ -403,6 +403,12 @@ macro_rules! functionality {
                 let sig = sk2.try_sign_with_rng(&mut rng, &message, b"context", true).unwrap();
                 let result = pk2.verify(&message, &sig, b"context");
                 assert!(result, "Signature failed to verify");
+
+                let (pk3, sk3) = KG::keygen_with_seeds(&[0u8; N], &[1u8; N], &[2u8; N]);
+                let sig = sk3.try_sign_with_rng(&mut rng, &message, b"context", true).unwrap();
+                let result = pk3.verify(&message, &sig, b"context");
+
+                assert!(result, "Signature failed to verify");
                 let result = pk2.verify(&message, &sig, b"some other context");
                 assert!(!result, "Signature should not have verified");
                 for ph in [Ph::SHA256, Ph::SHA512, Ph::SHAKE128, Ph::SHAKE256] {
@@ -420,9 +426,10 @@ macro_rules! functionality {
 }
 
 
-/// Functionality for the **SLH-DSA-SHA2-128s** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHA2-128s parameter set is claimed to be in security strength category 1.
+/// Functionality for the **SLH-DSA-SHA2-128s** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHA2-128s parameter set is claimed to be in security strength category 1.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_sha2_128s::try_keygen`] function below
 /// to generate both [`slh_dsa_sha2_128s::PublicKey`] and [`slh_dsa_sha2_128s::PrivateKey`] structs. The resulting
@@ -443,7 +450,8 @@ pub mod slh_dsa_sha2_128s {
     use crate::hashers::sha2_cat_1::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 16;
+    /// Seed size
+    pub const N: usize = 16;
     const H: usize = 63;
     const D: usize = 7;
     const HP: usize = 9;
@@ -468,9 +476,10 @@ pub mod slh_dsa_sha2_128s {
 }
 
 
-/// Functionality for the **SLH-DSA-SHAKE-128s** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHAKE-128s parameter set is claimed to be in security strength category 1.
+/// Functionality for the **SLH-DSA-SHAKE-128s** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHAKE-128s parameter set is claimed to be in security strength category 1.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_shake_128s::try_keygen`] function below
 /// to generate both [`slh_dsa_shake_128s::PublicKey`] and [`slh_dsa_shake_128s::PrivateKey`] structs. The resulting
@@ -491,7 +500,8 @@ pub mod slh_dsa_shake_128s {
     use crate::hashers::shake::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 16;
+    /// Seed size
+    pub const N: usize = 16;
     const H: usize = 63;
     const D: usize = 7;
     const HP: usize = 9;
@@ -516,9 +526,10 @@ pub mod slh_dsa_shake_128s {
 }
 
 
-/// Functionality for the **SLH-DSA-SHA2-128f** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHA2-128f parameter set is claimed to be in security strength category 1.
+/// Functionality for the **SLH-DSA-SHA2-128f** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHA2-128f parameter set is claimed to be in security strength category 1.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_sha2_128f::try_keygen`] function below
 /// to generate both [`slh_dsa_sha2_128f::PublicKey`] and [`slh_dsa_sha2_128f::PrivateKey`] structs. The resulting
@@ -539,7 +550,8 @@ pub mod slh_dsa_sha2_128f {
     use crate::hashers::sha2_cat_1::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 16;
+    /// Seed size
+    pub const N: usize = 16;
     const H: usize = 66;
     const D: usize = 22;
     const HP: usize = 3;
@@ -564,9 +576,10 @@ pub mod slh_dsa_sha2_128f {
 }
 
 
-/// Functionality for the **SLH-DSA-SHAKE-128f** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHAKE-128f parameter set is claimed to be in security strength category 1.
+/// Functionality for the **SLH-DSA-SHAKE-128f** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHAKE-128f parameter set is claimed to be in security strength category 1.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_shake_128f::try_keygen`] function below
 /// to generate both [`slh_dsa_shake_128f::PublicKey`] and [`slh_dsa_shake_128f::PrivateKey`] structs. The resulting
@@ -587,7 +600,8 @@ pub mod slh_dsa_shake_128f {
     use crate::hashers::shake::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 16;
+    /// Seed size
+    pub const N: usize = 16;
     const H: usize = 66;
     const D: usize = 22;
     const HP: usize = 3;
@@ -612,9 +626,10 @@ pub mod slh_dsa_shake_128f {
 }
 
 
-/// Functionality for the **SLH-DSA-SHA2-192s** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHA2-192s parameter set is claimed to be in security strength category 3.
+/// Functionality for the **SLH-DSA-SHA2-192s** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHA2-192s parameter set is claimed to be in security strength category 3.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_sha2_192s::try_keygen`] function below
 /// to generate both [`slh_dsa_sha2_192s::PublicKey`] and [`slh_dsa_sha2_192s::PrivateKey`] structs. The resulting
@@ -635,7 +650,8 @@ pub mod slh_dsa_sha2_192s {
     use crate::hashers::sha2_cat_3_5::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 24;
+    /// Seed size
+    pub const N: usize = 24;
     const H: usize = 63;
     const D: usize = 7;
     const HP: usize = 9;
@@ -660,9 +676,10 @@ pub mod slh_dsa_sha2_192s {
 }
 
 
-/// Functionality for the **SLH-DSA-SHAKE-192s** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHAKE-192s parameter set is claimed to be in security strength category 3.
+/// Functionality for the **SLH-DSA-SHAKE-192s** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHAKE-192s parameter set is claimed to be in security strength category 3.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_shake_192s::try_keygen`] function below
 /// to generate both [`slh_dsa_shake_192s::PublicKey`] and [`slh_dsa_shake_192s::PrivateKey`] structs. The resulting
@@ -683,7 +700,8 @@ pub mod slh_dsa_shake_192s {
     use crate::hashers::shake::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 24;
+    /// Seed size
+    pub const N: usize = 24;
     const H: usize = 63;
     const D: usize = 7;
     const HP: usize = 9;
@@ -708,9 +726,10 @@ pub mod slh_dsa_shake_192s {
 }
 
 
-/// Functionality for the **SLH-DSA-SHA2-192f** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHA2-192f parameter set is claimed to be in security strength category 3.
+/// Functionality for the **SLH-DSA-SHA2-192f** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHA2-192f parameter set is claimed to be in security strength category 3.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_sha2_192f::try_keygen`] function below
 /// to generate both [`slh_dsa_sha2_192f::PublicKey`] and [`slh_dsa_sha2_192f::PrivateKey`] structs. The resulting
@@ -731,7 +750,8 @@ pub mod slh_dsa_sha2_192f {
     use crate::hashers::sha2_cat_3_5::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 24;
+    /// Seed size
+    pub const N: usize = 24;
     const H: usize = 66;
     const D: usize = 22;
     const HP: usize = 3;
@@ -756,9 +776,10 @@ pub mod slh_dsa_sha2_192f {
 }
 
 
-/// Functionality for the **SLH-DSA-SHAKE-192f** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHAKE-192f parameter set is claimed to be in security strength category 3.
+/// Functionality for the **SLH-DSA-SHAKE-192f** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHAKE-192f parameter set is claimed to be in security strength category 3.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_shake_192f::try_keygen`] function below
 /// to generate both [`slh_dsa_shake_192f::PublicKey`] and [`slh_dsa_shake_192f::PrivateKey`] structs. The resulting
@@ -779,7 +800,8 @@ pub mod slh_dsa_shake_192f {
     use crate::hashers::shake::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 24;
+    /// Seed size
+    pub const N: usize = 24;
     const H: usize = 66;
     const D: usize = 22;
     const HP: usize = 3;
@@ -804,9 +826,10 @@ pub mod slh_dsa_shake_192f {
 }
 
 
-/// Functionality for the **SLH-DSA-SHA2-256s** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHA2-256s parameter set is claimed to be in security strength category 5.
+/// Functionality for the **SLH-DSA-SHA2-256s** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHA2-256s parameter set is claimed to be in security strength category 5.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_sha2_256s::try_keygen`] function below
 /// to generate both [`slh_dsa_sha2_256s::PublicKey`] and [`slh_dsa_sha2_256s::PrivateKey`] structs. The resulting
@@ -827,7 +850,8 @@ pub mod slh_dsa_sha2_256s {
     use crate::hashers::sha2_cat_3_5::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 32;
+    /// Seed size
+    pub const N: usize = 32;
     const H: usize = 64;
     const D: usize = 8;
     const HP: usize = 8;
@@ -852,9 +876,10 @@ pub mod slh_dsa_sha2_256s {
 }
 
 
-/// Functionality for the **SLH-DSA-SHAKE-256s** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHAKE_256s parameter set is claimed to be in security strength category 5.
+/// Functionality for the **SLH-DSA-SHAKE-256s** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHAKE_256s parameter set is claimed to be in security strength category 5.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_shake_256s::try_keygen`] function below
 /// to generate both [`slh_dsa_shake_256s::PublicKey`] and [`slh_dsa_shake_256s::PrivateKey`] structs. The resulting
@@ -875,7 +900,8 @@ pub mod slh_dsa_shake_256s {
     use crate::hashers::shake::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 32;
+    /// Seed size
+    pub const N: usize = 32;
     const H: usize = 64;
     const D: usize = 8;
     const HP: usize = 8;
@@ -900,9 +926,10 @@ pub mod slh_dsa_shake_256s {
 }
 
 
-/// Functionality for the **SLH-DSA-SHA2-256f** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHA2-256f parameter set is claimed to be in security strength category 5.
+/// Functionality for the **SLH-DSA-SHA2-256f** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHA2-256f parameter set is claimed to be in security strength category 5.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_sha2_256f::try_keygen`] function below
 /// to generate both [`slh_dsa_sha2_256f::PublicKey`] and [`slh_dsa_sha2_256f::PrivateKey`] structs. The resulting
@@ -923,7 +950,8 @@ pub mod slh_dsa_sha2_256f {
     use crate::hashers::sha2_cat_3_5::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 32;
+    /// Seed size
+    pub const N: usize = 32;
     const H: usize = 68;
     const D: usize = 17;
     const HP: usize = 4;
@@ -948,9 +976,10 @@ pub mod slh_dsa_sha2_256f {
 }
 
 
-/// Functionality for the **SLH-DSA-SHAKE-256f** security parameter set per FIPS 205 section 11. This includes specific
-/// sizes for the public key, secret key, and signature along with a number of internal constants. The
-/// SLH-DSA-SHAKE-256f parameter set is claimed to be in security strength category 5.
+/// Functionality for the **SLH-DSA-SHAKE-256f** security parameter set per FIPS 205 section 11.
+///
+/// This includes specific sizes for the public key, secret key, and signature along with a number of internal
+/// constants. The SLH-DSA-SHAKE-256f parameter set is claimed to be in security strength category 5.
 ///
 /// **1)** The basic usage is for an originator to start with the [`slh_dsa_shake_256f::try_keygen`] function below
 /// to generate both [`slh_dsa_shake_256f::PublicKey`] and [`slh_dsa_shake_256f::PrivateKey`] structs. The resulting
@@ -971,7 +1000,8 @@ pub mod slh_dsa_shake_256f {
     use crate::hashers::shake::{f, h, h_msg, prf, prf_msg, t_l};
     use crate::hashers::Hashers;
 
-    const N: usize = 32;
+    /// Seed size
+    pub const N: usize = 32;
     const H: usize = 68;
     const D: usize = 17;
     const HP: usize = 4;