+++ src/traits.rs
@@ -1,4 +1,6 @@
 use rand_core::CryptoRngCore;
+use rand_core::RngCore;
+use rand_core::CryptoRng;
 
 use crate::Ph;
 #[cfg(feature = "default-rng")]
@@ -99,7 +101,74 @@ pub trait KeyGen {
     fn try_keygen_with_rng(
         rng: &mut impl CryptoRngCore,
     ) -> Result<(Self::PublicKey, Self::PrivateKey), &'static str>;
+
+    /// Generates a public and private key pair specific to this security parameter set.
+    /// This function utilizes **three provided seeds** rather than a random number
+    /// generator in order to deterministically generate keys. This function operates
+    /// in constant-time relative to secret data.
+    /// # Errors
+    /// Returns an error when the random number generator fails.
+    /// # Examples
+    /// ```rust
+    /// use fips205::slh_dsa_shake_128s; // Could use any of the twelve security parameter sets.
+    /// use fips205::traits::{KeyGen, SerDes, Signer, Verifier};
+    /// # use std::error::Error;
+    /// # use rand_core::OsRng;
+    /// #
+    /// # fn main() -> Result<(), Box<dyn Error>> {
+    ///
+    /// let msg_bytes = [0u8, 1, 2, 3, 4, 5, 6, 7];
+    /// let mut rng = OsRng;
+    ///
+    /// // Generate both public and secret keys. This only fails when the provided rng fails.
+    /// let (pk1, sk) = slh_dsa_shake_128s::KG::keygen_with_seeds(&[0u8; slh_dsa_shake_128s::N],
+    ///                 &[1u8; slh_dsa_shake_128s::N], &[1u8; slh_dsa_shake_128s::N]);
+    /// // Use the secret key to generate a signature. The second parameter is the
+    /// // context string (often just an empty &[]), and the last parameter selects
+    /// // the preferred hedged variant. This only fails when the OS rng fails.
+    /// let sig_bytes = sk.try_sign(&msg_bytes, b"context", true)?;
+    ///
+    ///
+    /// // Serialize the public key, and send with message and signature bytes. These
+    /// // statements model sending byte arrays over the wire.
+    /// let (pk_send, msg_send, sig_send) = (pk1.into_bytes(), msg_bytes, sig_bytes);
+    /// let (pk_recv, msg_recv, sig_recv) = (pk_send, msg_send, sig_send);
+    ///
+    ///
+    /// // Deserialize the public key. This only fails on a malformed key.
+    /// let pk2 = slh_dsa_shake_128s::PublicKey::try_from_bytes(&pk_recv)?;
+    /// // Use the public key to verify the msg signature
+    /// let v = pk2.verify(&msg_recv, &sig_recv, b"context");
+    /// assert!(v);
+    /// # Ok(())
+    /// # }
+    /// ```
+    fn keygen_with_seeds<const N: usize>(
+        sk_seed: &[u8; N], sk_prf: &[u8; N], pk_seed: &[u8; N]
+    ) -> (Self::PublicKey, Self::PrivateKey) {
+        Self::try_keygen_with_rng(&mut DummyRng {data: [*sk_seed, *sk_prf, *pk_seed], i: 0 }).expect("rng will not fail")
+    }
+
 }
+
+// This is for the deterministic keygen functions; will be refactored more nicely
+struct DummyRng<const N: usize> { data: [[u8; N]; 3], i: usize }
+
+impl<const N: usize> RngCore for DummyRng<N> {
+    fn next_u32(&mut self) -> u32 { unimplemented!() }
+
+    fn next_u64(&mut self) -> u64 { unimplemented!() }
+
+    fn fill_bytes(&mut self, _out: &mut [u8]) { unimplemented!() }
+
+    fn try_fill_bytes(&mut self, out: &mut [u8]) -> Result<(), rand_core::Error> {
+        out.copy_from_slice(&self.data[self.i]);
+        self.i += 1;
+        Ok(())
+    }
+}
+
+impl<const N: usize> CryptoRng for DummyRng<N> {}
 
 
 /// The Signer trait is implemented for the `PrivateKey` struct on each of the security parameter sets