+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.1 (2024-12-22)
+
+- Added keygen with seeds for deterministic keygen
+- Added two fuzzing harnesses (still some work to go)
+
+## 0.4.0 (2024-10-04)
+
+- Updated to FIPS 205 final spec
+
 ## 0.1.2 (2024-03-15)
 
 - Internal improvements, removed dependency on generic-array, MSRV at 1.70